#6 controller에 있던 논리 로직들 service로 분리
