+++ src/posts/posts.service.ts
@@ -1,4 +1,113 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
+
+/**
+ * author: string;
+ * title: string;
+ * content: string;
+ * likeCount: number;
+ * commentCount: number;
+ */
+
+export interface PostModel {
+  id: number;
+  author: string;
+  title: string;
+  content: string;
+  likeCount: number;
+  commentCount: number;
+}
+
+let posts: PostModel[] = [
+  {
+    id: 1,
+    author: 'teddy',
+    title: 'bear',
+    content: 'dad',
+    likeCount: 1111,
+    commentCount: 123,
+  },
+  {
+    id: 2,
+    author: 'teddy2',
+    title: 'bear2',
+    content: 'dad2',
+    likeCount: 1111,
+    commentCount: 123,
+  },
+  {
+    id: 3,
+    author: 'teddy3',
+    title: 'bear3',
+    content: 'dad3',
+    likeCount: 1111,
+    commentCount: 123,
+  },
+];
 
 @Injectable()
-export class PostsService {}
+export class PostsService {
+  getAllPosts() {
+    return posts;
+  }
+
+  getPostById(id: number) {
+    const post = posts.find((post) => post.id === +id);
+
+    if (!post) {
+      throw new NotFoundException();
+    }
+
+    return post;
+  }
+
+  createPost(author: string, title: string, content: string) {
+    const post: PostModel = {
+      id: posts[posts.length - 1].id + 1,
+      author,
+      title,
+      content,
+      likeCount: 0,
+      commentCount: 0,
+    };
+
+    posts = [...posts, post];
+
+    return post;
+  }
+
+  updatePost(postId: number, author: string, title: string, content: string) {
+    const post = posts.find((post) => post.id === postId);
+
+    if (!post) {
+      throw new NotFoundException();
+    }
+
+    if (author) {
+      post.author = author;
+    }
+
+    if (title) {
+      post.title = title;
+    }
+
+    if (content) {
+      post.content = content;
+    }
+
+    posts = posts.map((prevPost) => (prevPost.id === postId ? post : prevPost));
+
+    return post;
+  }
+
+  deletePost(postId: number) {
+    const post = posts.find((post) => post.id === postId);
+
+    if (!post) {
+      throw new NotFoundException();
+    }
+
+    posts = posts.filter((post) => post.id !== postId);
+
+    return postId;
+  }
+}
